New files and functionalities added
The aim of this commit is to make a better structure in order to test
import functionalities of sass.

More preciselly: here is added many files which are all included in
general/main.scss.
In Gruntfile.js it's specified path general/main.scss which is the only
file which will be built through grunt compass command.